+++ tests/testthat/test-loo-loso.R
@@ -77,6 +77,8 @@ test_that("loo object structure and internal identities", {
       "log_cpo_1",
       "log_cpo_2",
       "det_term",
+      "k_max",
+      "k_sum",
       "ok"
     )
   )